requirements/development.txt
amqp==5.1.1
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.4
beautifulsoup4==4.12.2
billiard==3.6.4.0
boto3==1.26.12
bs4==0.0.1
CacheControl==0.12.11
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Collectfast==2.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
dj-database-url==1.0.0
Django==4.2.1
django-anymail==8.1
django-autoslug==1.9.8
django-axes==5.40.1
django-cleanup==7.0.0
django-cors-headers==3.13.0
django-dirtyfields==1.9.0
django-dotenv==1.4.2
django-extensions==3.2.1
django-filter==22.1
django-health-check==3.17.0
django-ipware==4.0.2
django-phonenumber-field==7.0.1
django-phonenumbers==1.0.1
django-redis==5.2.0
django-redis-sessions==0.6.2
django-rest-passwordreset==1.3.0
django-silk==5.0.4
django-simple-history==3.2.0
django-storages==1.13.2
django-versatileimagefield==2.2
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
elastic-transport==8.4.0
elasticsearch==8.5.0
exceptiongroup==1.1.1
firebase-admin==6.0.1
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
gprof2dot==2022.7.29
grpcio==1.54.0
grpcio-status==1.54.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
kombu==5.2.4
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
msgpack==1.0.5
numpy==1.24.3
outcome==1.2.0
packaging==23.1
pandas==1.5.3
phonenumbers==8.13.10
Pillow==9.3.0
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.22.3
psycopg2-binary==2.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.0
pycparser==2.21
Pygments==2.16.1
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==0.8.11
python-magic==0.4.27
pytz==2021.3
PyYAML==6.0
redis==4.5.4
requests==2.28.2
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
selenium==4.7.2
simplejson==3.19.1
six==1.15.0
sniffio==1.3.0
sorl-thumbnail==12.9.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.3
tomli==2.0.1
trio==0.22.0
trio-websocket==0.10.2
twilio==7.16.4
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
wsproto==1.2.0
